update functions to remove test errors

+++ project/searching.py
@@ -16,9 +16,9 @@ def binary_search(arr, target):
     high = len(arr)-1
 
     while low <= high:
-        middle = (low+high)/2
+        middle = int((low+high)/2)
         if target < arr[middle]:
-            high = middle -1
+            high = middle-1
         elif target > arr[middle]:
             low = middle+1
         else:
@@ -30,7 +30,7 @@ def binary_search(arr, target):
 # STRETCH: write a recursive implementation of Binary Search
 def binary_search_recursive(arr, target, low, high):
 
-    middle = (low+high)/2
+    middle = int((low+high)/2)
 
     if len(arr) == 0:
         return -1
